payment-links: move list components to List
