+++ packages/pilot/stories/containers/PaymentLinks/List/Filter/index.js
@@ -3,27 +3,28 @@ import { path, split } from 'ramda'
 import moment from 'moment'
 
 import { action } from '@storybook/addon-actions'
-import Section from '../../../Section'
-import Filter from '../../../../src/containers/PaymentLinks/Filter'
+import Section from '../../../../Section'
+import Filter from '../../../../../src/containers/PaymentLinks/List/Filter'
 
-import translations from '../../../../public/locales/pt/translations.json'
+import translations from '../../../../../public/locales/pt/translations.json'
 
 const t = sentence => path(split('.', sentence), translations)
 
 const FilterExample = () => (
   <Section>
     <Filter
-      searchData={{
+      query={{
         active: true,
         dates: {
           end: moment(),
           start: moment().subtract(1, 'month'),
         },
         inactive: false,
-        linkNameQuery: 'Pulseira gótica',
+        name: 'Pulseira gótica',
       }}
-      onReset={action('onReset')}
       onApply={action('onApply')}
+      onQueryChange={action('onQueryChange')}
+      onReset={action('onReset')}
       t={t}
     />
   </Section>
