+++ packages/pilot/stories/containers/PaymentLinks/index.js
@@ -1,13 +1,7 @@
-import Filter from './Filter'
 import Details from './Details'
-import NewLinksCard from './NewLinksCard'
-import PaymentLinksList from './PaymentLinksList'
-import PaymentLinkAdd from './PaymentLinkAdd'
+import List from './List'
 
 export default {
   Details,
-  Filter,
-  NewLinksCard,
-  PaymentLinkAdd,
-  PaymentLinksList,
+  List,
 }
