+++ packages/pilot/stories/containers/index.js
@@ -452,23 +452,23 @@ storiesOf('Containers|Empty State', module)
   .add('Default', () => <EmptyState.EmptyStateDefault />)
   .add('MDRzao', () => <EmptyState.EmptyStateMDRzao />)
 
-storiesOf('Containers|Payment Links/Filter', module)
-  .add('Filter', () => <PaymentLinks.Filter />)
+storiesOf('Containers|Payment Links/List/Filter', module)
+  .add('Filter', () => <PaymentLinks.List.Filter />)
 
-storiesOf('Containers|Payment Links/NewLinksCard', module)
-  .add('NewLinksCard', () => <PaymentLinks.NewLinksCard />)
+storiesOf('Containers|Payment Links/List/NewLinksCard', module)
+  .add('NewLinksCard', () => <PaymentLinks.List.NewLinksCard />)
 
-storiesOf('Containers|Payment Links/PaymentLinkList', module)
-  .add('List', () => <PaymentLinks.PaymentLinksList.WithList />)
-  .add('List loading', () => <PaymentLinks.PaymentLinksList.WithListLoading />)
-  .add('Empty list', () => <PaymentLinks.PaymentLinksList.WithEmptyList />)
+storiesOf('Containers|Payment Links/List/PaymentLinkList', module)
+  .add('List', () => <PaymentLinks.List.PaymentLinksList.WithList />)
+  .add('List loading', () => <PaymentLinks.List.PaymentLinksList.WithListLoading />)
+  .add('Empty list', () => <PaymentLinks.List.PaymentLinksList.WithEmptyList />)
 
-storiesOf('Containers|Payment Links/PaymentLinkAdd', module)
-  .add('AddModal', () => <PaymentLinks.PaymentLinkAdd.PaymentLinkAddModal />)
-  .add('FirstStep', () => <PaymentLinks.PaymentLinkAdd.PaymentLinkFirstStep />)
-  .add('SecondStep', () => <PaymentLinks.PaymentLinkAdd.PaymentLinkSecondStep />)
-  .add('SuccessStep', () => <PaymentLinks.PaymentLinkAdd.PaymentLinkSuccessStep />)
-  .add('ErrorStep', () => <PaymentLinks.PaymentLinkAdd.PaymentLinkErrorStep />)
+storiesOf('Containers|Payment Links/List/PaymentLinkAdd', module)
+  .add('AddModal', () => <PaymentLinks.List.PaymentLinkAdd.PaymentLinkAddModal />)
+  .add('FirstStep', () => <PaymentLinks.List.PaymentLinkAdd.PaymentLinkFirstStep />)
+  .add('SecondStep', () => <PaymentLinks.List.PaymentLinkAdd.PaymentLinkSecondStep />)
+  .add('SuccessStep', () => <PaymentLinks.List.PaymentLinkAdd.PaymentLinkSuccessStep />)
+  .add('ErrorStep', () => <PaymentLinks.List.PaymentLinkAdd.PaymentLinkErrorStep />)
 
 storiesOf('Containers|Payment Links/Details/LinkStatement', module)
   .add('Default', () => <PaymentLinks.Details.LinkStatement.LinkStatementDefault />)
